+++ src/lib/IExecTaskModule.js
@@ -1,6 +1,10 @@
 import IExecModule from './IExecModule.js';
 import { show, obsTask, claim } from '../common/execution/task.js';
 import { fetchTaskResults } from '../common/execution/result.js';
+import {
+  fetchAllReplicatesLogs,
+  fetchTaskOffchainInfo,
+} from '../common/execution/debug.js';
 
 export default class IExecTaskModule extends IExecModule {
   constructor(...args) {
@@ -18,5 +22,12 @@ export default class IExecTaskModule extends IExecModule {
       fetchTaskResults(await this.config.resolveContractsClient(), taskid, {
         ipfsGatewayURL: await this.config.resolveIpfsGatewayURL(),
       });
+    this.fetchLogs = async (taskid) =>
+      fetchAllReplicatesLogs(
+        await this.config.resolveContractsClient(),
+        taskid,
+      );
+    this.fetchOffchainInfo = async (taskid) =>
+      fetchTaskOffchainInfo(await this.config.resolveContractsClient(), taskid);
   }
 }